project_website/requirements.txt
psycopg2==2.9.9
django-stubs==4.2.7
# types-psycopg2==2.9.21.20240106
scikit-learn==1.3.2
# matplotlib==3.4.3
# ChatterBotPy == 1.1.3
SQLAlchemy==2.0.25
pyyaml==5.3
langchain==0.1.2
langchain-community==0.0.14
langchain-openai==0.0.3
langchain-experimental==0.0.49